[feat] : 홈 화면과 나의 학습 페이지 기능 합치기
